+++ services/registries.js
@@ -204,14 +204,12 @@ export class RegistryService extends AbstractService {
 
     // TODO: check this
     indexes.push([{ _keywordsLabels: 1 }, {}])
-    indexes.push([{ _keywordsPublisher: 1 }, {}])
     indexes.push([{ "_keywordsLabels.0": 1 }, {}])
     indexes.push([
       {
         _keywordsNotation: "text",
         _keywordsLabels: "text",
         _keywordsOther: "text",
-        _keywordsPublisher: "text",
       },
       {
         name: "text",
@@ -220,7 +218,6 @@ export class RegistryService extends AbstractService {
           _keywordsNotation: 10,
           _keywordsLabels: 6,
           _keywordsOther: 3,
-          _keywordsPublisher: 3,
         },
       },
     ])